Handle possible presence of aiodns on Windows
aiodns on Windows requires SelectorEventLoop, which does not support asyncio subprocess, breaking the "execute" plugin. It gets installed due to being specified as mandatory requirement of slixmpp.

aiohttp, which also uses aiodns as an optional requirement, used a workaround to prevent it from failing on Windows, however, it got removed in version 3.10.0

This commit bundles fixes for this situation:

- monkey-patch slixmpp.xmlstream.resolver in msg2jbr_slixmpp.py
- limit aiohttp version to <3.10.0
- remove aiodns from github action for building .exe before running pyinstaller