Clear MessageBus._subscriptions on (re)start
Proper solution would be to make it an instance variable and have a single MessageBus instance passed to __init__ of actors and chains. It would open possibility for implementing running multiple independent config files. However, it requires updating __init__ signature for every actor, which is too much, so it stays like this for now.